numPath first succesful pass
It uses Dijkstra's algorithm to find the shortest path from dst to all other vertices.

it then uses BST to add ages to a tree if the next vertex has a smaller distance to dst than the current vertex. if the next vertex is equal to dst, then the number of paths is incremented.